mysite/requirement.txt
appdirs==1.4.3
asgiref==3.6.0
attrs==19.3.0
Babel==2.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cliff==3.1.0
cloud-init==21.4
cmd2==0.8.5
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debtcollector==2.0.0
decorator==4.4.2
defer==1.0.6
deprecation==2.0.7
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.5
docutils==0.16
dogpile.cache==0.9.0
entrypoints==0.3
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
iso8601==0.1.12
jenkinsapi==0.3.11
Jinja2==2.10.1
jmespath==0.9.4
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
keystoneauth1==4.0.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
msgpack==0.6.2
multi-key-dict==2.0.3
munch==2.3.2
mysqlclient==2.1.1
netaddr==0.7.19
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openstacksdk==0.46.0
os-client-config==1.31.2
os-service-types==1.7.0
osc-lib==2.0.0
oslo.config==8.0.2
oslo.context==3.0.2
oslo.i18n==4.0.1
oslo.log==4.1.1
oslo.serialization==3.1.1
oslo.utils==4.1.1
packaging==20.3
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
prettytable==0.7.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.6
pyperclip==1.7.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-cinderclient==7.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-jenkins==1.7.0
python-keystoneclient==4.0.0
python-novaclient==17.0.0
python-openstackclient==5.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
requestsexceptions==1.4.0
rfc3986==1.3.2
roman==2.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.4.3
ssh-import-id==5.10
stevedore==1.32.0
systemd-python==234
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
wrapt==1.11.2
xkit==0.0.0
zipp==1.0.0
